Add SL as codeowners for snaps-owned controllers
Whenever the version of a package in this monorepo is bumped, all
dependencies on that package in other workspace packages are also
bumped. This happens frequently in release PRs.

The Snaps team is listed as codeowners of `permission-controller`,
`notification-controller`, and `rate-limit-controller`, so when any
changes to these packages occur in a PR, the Snaps team must approve
that PR. This means that if a dependency on a package is bumped in any
of these controllers within a release PR, then the Snaps team must
approve that release — even if none of those controllers are included in
the release. This creates an obstacle to creating releases quickly.

To fix this, this commit adds the Shared Libraries team as codeowners of
these controllers as well. This way, they can ask the Snaps team to
approve only if it is necessary.