refactor(queued-request-controller): Migrate request count to state (#3919)
## Explanation

The current number of requests queued by the `QueuedRequestController`
is currently stored as a private instance variable and exposed by the
`length` method and the `countChanged` event. This works, but the
conventional way of tracking and sharing state is to use controller
state. This is what controller state is for, and more broadly this is
why controllers exist: to manage state.

To better align with our conventions, the requests queue length has been
migrated into controller state. It is now represented by the state
property `queuedRequestCount`.

The old `length` method and `countChanged` event remains, but they have
been marked as deprecated. They can be removed in a future release.

Additionally, types have been added for the built-in `stateChanged`
event and `getState` action.


## References

This was done to help unblock RPC request queue batching (#3763). This
isn't strictly required for that work, but it made it simpler to
implement.

## Changelog

### `@metamask/queued-request-controller

#### Added
- Add `queuedRequestCount` state
- Add pre-existing `getState` action and `stateChange` event to type
exports

#### Changed
- Deprecate the `length` method in favor of the `queuedRequestCount`
state
- Deprecate the `countChanged` event in favor of the `stateChange` event

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate