refactor(queued-request-controller): Refactor request processing
The `QueuedRequestMiddleware` and `QueuedRequestController` work
together to determine when to queue or execute a request, and whether
to trigger a network switch. Previously the network switch step was
performed by the middleware. It has been refactored to happen inside
the controller instead. This change results in various breaking changes
to both the controller and middleware, but when used together they
should behave exactly the same as before.

Relates to #3763