Add workflow to check presence of DO-NOT-MERGE
The `DO-NOT-MERGE` label is designed to prevent other people from
merging a PR (perhaps because it is blocking another PR or for some
other reason). In reality, however, this is merely an alert, and it is
still possible for a PR which has this label to be merged.

This commit adds a GitHub workflow which runs when opening a PR or
changing the labels for a PR and fails if the PR has a `DO-NOT-MERGE`
label. Once this workflow is in place, the
`ensure-block-pr-labels-absent` check can be added to the branch
protection settings to truly enforce that a PR cannot be merged while
the label is in place.