Add workflow to check presence of DO-NOT-MERGE (#3960)
The `DO-NOT-MERGE` label is designed to prevent other people from
merging a PR (perhaps because it is blocking another PR or for some
other reason). In reality, however, this is merely an alert, and it is
still possible for a PR which has this label to be merged.

This commit adds a GitHub workflow which runs when opening a PR or
changing the labels for a PR and fails if the PR has a `DO-NOT-MERGE`
label. Once this workflow is in place, the
`ensure-block-pr-labels-absent` check can be added to the branch
protection settings to truly enforce that a PR cannot be merged while
the label is in place.

---

The extension has a similar check:
https://github.com/MetaMask/metamask-extension/blob/275523a24e4f496384943f107369810cd93aa86e/.github/workflows/check-pr-labels.yml.
I simplified the workflow.