Bump `@metamask/rpc-errors` to `^6.2.1` (#3970)
## Explanation

This bumps `@metamask/rpc-errors` to `^6.2.1`. This is a more permanent
fix for a type issue (initially solved by #3954).

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/approval-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/assets-controllers`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/json-rpc-engine`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/network-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/permission-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/queued-request-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/rate-limit-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/signature-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/transaction-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

### `@metamask/user-operation-controller`

- **CHANGED**: Bump `@metamask/rpc-errors` from `^6.2.0` to `^6.2.1`.

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate