Simplify error handling
The middleware callback had previously been setup as a secondary way
to return an error from the controller to the middleware. This was not
actually used though, and not necessary even in future PRs, so it has
been removed.

Any errors encountered in the controller are returned to the middleware
by being thrown from `enqueueRequest`.