refactor(queued-request-controller): Refactor request processing (#3968)
## Explanation

The `QueuedRequestMiddleware` and `QueuedRequestController` work
together to determine when to queue or execute a request, and whether to
trigger a network switch. Previously the network switch step was
performed by the middleware. It has been refactored to happen inside the
controller instead. This change results in various breaking changes to
both the controller and middleware, but when used together they should
behave exactly the same as before.

This change was made as part of the effort to refactor the RPC queue to
batch requests by origin (#3763). It was easier to accomplish that if
the network switch step happened inside the controller, because the
controller would have additional state needed to make that decision for
batches rather than individual requests.

## References

Relates to #3763

## Changelog

### `@metamask/queued-request-controller`

#### Changed
- **BREAKING:** The `QueuedRequestController` method `enqueueRequest` is
now responsible for switching the network before processing a request,
rather than the `QueuedRequestMiddleware`
- Functionally the behavior is the same: before processing each request,
we compare the request network client with the current selected network
client, and we switch the current selected network client if necessary.
- The `QueuedRequestController` messenger has four additional allowed
actions:
    - `NetworkController:getState`
    - `NetworkController:setActiveNetwork`
    - `NetworkController:getNetworkConfigurationByNetworkClientId`
    - `ApprovalController:addRequest`
- The `QueuedRequestController` method `enqueueRequest` now takes one
additional parameter, the request object
- The `QueuedRequestMiddleware` no longer has a controller messenger.
Instead it takes the `enqueueRequest` method as a parameter.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: jiexi <jiexiluan@gmail.com>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>