Check labels for a PR when it is updated (#4005)
Updating a branch against its base branch from the GitHub UI seems to
reset the status of the `ensure-blocking-pr-labels-absent` check, but
since it only runs when a PR is opened or when labels are added or
removed, it won't get run again, keeping the PR in limbo. Adding
`synchronize` to the list of actions fixes this.