refactor(controller-utils): replace `any` with types for type safety (#3975)
## Explanation

The current implementation uses any type for parameters in certain
utility functions within controller-utils, which lacks type safety and
can lead to runtime errors if incorrect values are passed. This approach
needs to change to ensure the reliability and maintainability of the
codebase.

## References

- Fixes #3717

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/controller-utils`

#### Fixed

- **BREAKING**: Replaced any type with BN for BNToHex and fractionBN
functions to enhance type safety.
- **BREAKING**: Replaced any type with unknown in logOrRethrowError.
- **BREAKING**: Replaced any type with string in isNetworkType.


## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate