Correct type of `selector` arg to `subscribe` (#4012)
The ControllerMessenger and RestrictedControllerMessenger both have a
`subscribe` method which takes an optional third argument, a selector
function. The type of this function, however, is wrong, as the types of
its arguments are always inferred as `never`. This results in the
following:

```
messenger.subscribe(
  // eventType
  'CountController:stateChange',
  // handler
  (state) => {
    // ...
  },
  // selector
  (state) => {    // NOTE: The type of `state` here will be `never`
    // ...
  }
)
```

The arguments are `never` because `SelectorFunction` is being used
incorrectly. This utility type is supposed to take an event type and a
return type for the selector type, and then it will infer the arguments
for the resulting function using `ExtractEventPayload`. However, in
defining the type of `selector`, `subscribe` uses `ExtractEventPayload`
ahead of time, so it ends up passing a function type to
`SelectorFunction`.

This incorrect use of `SelectorFunction` may be due to the fact that
this utility type does not take an `EventType`. This commit adds
`EventType` to `SelectorFunction` as a parameter so that we
can use it to declare the correct type for `selector`.

The `EventSubscriptionMap` type also uses `SelectorFunction`
incorrectly. While this does not cause any issues, this commit also
corrects this as well.