Initialize controller state as deeply frozen (#4011)
It should not be possible to mutate state outside of a BaseController v2
controller's `update` method, because the `stateChange` event for that
controller will not be fired and so it can lead to surprises in other
parts of the system which make use of that state. However, whereas
controller state is frozen after `update` is called thanks to Immer,
until that point, consumers are completely free to directly access it
and change it. For instance, this does not throw when it should:

```
const controller = new MyController();
controller.state.foo = "bar";
```

This is also possible:

```
const controller = new MyController();
controller.state.foos.push("bar");
```

This commit corrects this issue by using Immer to deep-freeze state
inside of the BaseController v2 constructor when it is set.