[queued-request-controller] Batch RPC requests (#3781)
## Explanation

The `QueuedRequestController` will now batch RPC requests by origin.
Each batch of requests will be processed in parallel, allowing existing
features relating to groups of requests to function correctly with
queuing enabled (e.g. the confirmation navigation buttons, and the
"Reject all" button). The batching by origin also ensures that
confirmations from different dapps are never displayed together in a
group, which will help prevent users from confusing which dapp each
confirmation is from.

The meaning of `queuedRequestCount` has changed; it now represents the
total number of queued requests, i.e. those that are queued but are
**not** being processed. Previously it included processing requests as
well. Given that the number of confirmations awaiting approval is
already tracked elsewhere, it was not useful to double count them by
including them in the queued request count as well.

Additionally, when the `QueuedRequestController` requires the globally
selected network to be changed, we now emit an event rather than
triggering a confirmation. This aligns with recent designs for this use
case.

The actions `NetworkController:getNetworkConfigurationByNetworkClientId`
and `ApprovalController:addRequest` were only used for triggering the
switch network confirmation, so they are no longer needed now. They have
been removed. This removes the last dependency of this controller on the
`ApprovalController`, so it has been removed as a dependency as well.

## References

Closes #3763
Fixes #3967
Closes #3983

## Changelog

### `@metamask/queued-request-controller`

#### Changed
- **BREAKING**: The `QueuedRequestController` will now batch queued
requests by origin
  - All of the requests in a single batch will be processed in parallel.
- Requests get processed in order of insertion, even across
origins/batches.
- All requests get processed even in the event of preceding requests
failing.
- **BREAKING:** The `queuedRequestCount` state no longer includes
requests that are currently being processed. It just counts requests
that are queued.
- **BREAKING:** The `QueuedRequestController` no longer triggers a
confirmation when a network switch is needed
  - The network switch now happens automatically, with no confirmation.
- A new `QueuedRequestController:networkSwitched` event has been added
to communicate when this has happened.
- The `QueuedRequestController` messenger no longer needs access to the
actions `NetworkController:getNetworkConfigurationByNetworkClientId` and
`ApprovalController:addRequest`.
- The package `@metamask/approval-controller` has been completely
removed as a dependency

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>
Co-authored-by: Alex Donesky <adonesky@gmail.com>