fix: Preview build doc corrections
The preview build documentation wrongly implied that the steps for
using a local or preview build were different for MetaMask team members
versus independent contributors. The steps are exactly the same.

Additionally, the examples that use npm have been removed. The override
syntax listed was not correct (npm doesn't use the same syntax for the
version specifier), and the other instructions all reference Yarn
anyway so the document already doesn't make much sense for a reader
only familiar with npm.