fix: include gas fees in user operations when using a paymaster (#4032)
## Explanation

### Description:
When a paymaster is used, the user operation still requires the
inclusion of gas fee values (`maxFeePerGas` and `maxPriorityFeePerGas`)
to enable the bundler and entrypoint to charge the paymaster
appropriately. This PR addresses an issue where the gas fees are
currently lost and submitted as empty to the bundler when using a
paymaster.

### Changes:
Detect when [getTransactionMetadata has set the gas fees to zero because
the userOperation has a
paymaster](https://github.com/MetaMask/core/blob/af668da3c09b62a43ae1f6c91bfd1a0f4241a610/packages/user-operation-controller/src/utils/transaction.ts#L86-L90)
and do not update the user operation gas fees.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->
* Fixes https://github.com/MetaMask/MetaMask-planning/issues/2180

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/user-operation-controller`

- **FIXED**: include gas fees in user operations when using a paymaster

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate