fix: Preview build doc corrections (#4053)
## Explanation

The preview build documentation wrongly implied that the steps for using
a local or preview build were different for MetaMask team members versus
independent contributors. The steps are exactly the same.

Additionally, the examples that use npm have been removed. The override
syntax listed was not correct (npm doesn't use the same syntax for the
version specifier), and the other instructions all reference Yarn anyway
so the document already doesn't make much sense for a reader only
familiar with npm.

## References

N/A

## Changelog

None

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate