Fix intermittently failing TransactionController test (#4057)
In this test, it seems that two `eth_sendRawTransaction` requests are
occurring when only one transaction was added. I am not 100% sure why
this is, but we've seen this type of problem in tests that use a
combination of HTTP requests mocked via Nock as well as the block
tracker, so my best guess is that there is some state which is not
cleaned up correctly and is therefore bleeding into other tests. The
solution to add an extra request mock is a bit of a band-aid, but
considering that we haven't seen this problem for any other test, it
should fix it for the time being.