fix: add updateNft fct (#4008)
## Explanation

When a user imports an NFT when the display NFT media toggle is off and
IPFS toggle is off, then enables the toggles, the image of the NFT will
not be displayed.

This PR adds a new fct that fetches the NFT metadata and udpates the
state. This fct will be called from extension based on the toggles and
if the name/description/image attributes are null

## References

* Related to extension PR
[#67890](https://github.com/MetaMask/metamask-extension/pull/23293)


## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **ADDED**: Added new fct updateNftMetadata that fetches nft metadata
and updates the state


## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate