fix: export const (#4079)
## Explanation

PR to export `SUPPORTED_CHAIN_IDS` constant.
This constant is imported in extension like this:
`import { SUPPORTED_CHAIN_IDS } from
'@metamask/assets-controllers/dist/token-prices-service/codefi-v2';`
Which was causing issues when trying to upgrade assets-controllers to
latest. It was not able to find it.
With this change we can import it like this instead of from dist which
can abstract how the code is transpiled;
`import { SUPPORTED_CHAIN_IDS } from '@metamask/assets-controllers';`

## References

* Related to
[#67890](https://github.com/MetaMask/metamask-extension/pull/23574)


## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **ADDED**: Added export for const SUPPORTED_CHAIN_IDS 

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate