Include errors in simulation data (#4067)
Export all simulation types. 
Include error data in the `SimulationData` type including `code` and
`message` if available, plus whether the error is due to the transaction
being reverted in the simulation.
Continue to trigger simulation when adding a transaction, but do not
block the creation of the approval request.
Automatically update the simulation data if the transaction parameters
are changed, specifically `to`, `data` and `value`.
Replace use of `JsonRpcProvider` with `fetch` to allow greater control
when parsing error responses.
Provide zero value gas fees in simulation requests to disable gas fee
suggestions.