Add workspaces:list-versions root script
From time to time, it's useful to have a list of the publishable
packages inside this monorepo and their current versions. Yarn does not
make this available via `yarn workspaces list`, so we have to do this
ourselves.

`yarn workspaces:list-versions` can be run to get a human-readable list,
or `yarn workspaces:list-versions --json` can be run to get a
machine-readable version.