Add guidelines for writing controllers
So far we have not have any documentation on how to write controllers
the BaseController v2 way. These guidelines aim to fill in the missing
pieces. Note that while each guideline has examples, we are still
missing a complete example controller file which demonstrates all of the
guidelines in one go. That will come in a future PR.