fix: use hardcoded Infura gas API urls (#4068)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

`legacyAPIEndpoint` and `EIP1559APIEndpoint` options are the gas API
urls used in `GasFeeController`.
This PR aims to remove those and put the new Infura URL as hardcoded. 

New Infura API also expects an auth header on requests, hence this PR is
adding new options `infuraAPIKey` to constructor.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

Fixes: https://github.com/MetaMask/MetaMask-planning/issues/2254

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/gas-fee-controller`

- **BREAKING**: Removed the constructor options `legacyAPIEndpoint` and
`EIP1559APIEndpoint`. These URLs are now hardcoded within the
controller.
- **BREAKING**: The controller's constructor now requires
`infuraAPIKey`. This is used to construct and send the Authorization
header for Infura gas API requests.

## Checklist

- [X] I've updated the test suite for new or updated code as appropriate
- [X] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [X] I've highlighted breaking changes using the "BREAKING" category
above as appropriate