Compare selector return values deeply
Currently, it is difficult to use the messenger to subscribe to a subset
of state changes by using the three-argument form of
`ControllerMessenger.subscribe` — that is, passing a selector function.
To be used effectively, the selector must return only one property in
state and the value of that property must be a primitive. If either of
these is not true — for instance, if the developer wishes to select more
than one property in state, and so has the selector return an array of
values rather than a single value — then the event listener will be
fired for every state change. This occurs because we perform a shallow
equality check when comparing the return value of the selector for the
previous state change and the current return value of the selector. This
commit replaces this with a deep equality check. The `just-compare`
package is being used in order to avoid a dependency on Lodash, which is
a rather large library, and add minimal size to the bundle instead.