Zb/queue flush event (#4139)
## Explanation

This PR is a follow up to
[this](https://github.com/MetaMask/core/pull/4064/files) - where instead
of calling the flush method directly after successfully handling a
request, we instead subscribe to SelectedNetworkController state changes
and flush the queue for the domain in question when the network changes.

One potential issue that may arise is relating to a possible race
condition. The question is when will the subscriptions fire - will they
happen before processing the next request starts? It's hard to say for
certain.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/request-queue-controller`

- **ADDED**: A method for flushing or clearing the queue of items for a
particular domain
- **FIXED**: calling `wallet_switchEthereumChain` will clear all queued
transactions from the domain, across all queued 'batches'

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate