Bump @spruceid/siwe-parser to 2.1.0 (#4141)
## Explanation
This pull request bumps the siwe-parser to the latest version which adds
additional support for the `scheme` parameter. I also refactor the tests
to explicitly call the siwe-parser library to act as a smoke test that
can catch any issues in the future. This will make breaking changes
easier to catch.

## References

https://github.com/MetaMask/MetaMask-planning/issues/2278

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

None

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate